`Int32` -> `Double`

+++ Packages/MicrophonePitchDetector/Sources/MicrophonePitchDetector/PitchTracker.swift
@@ -6,10 +6,10 @@ public final class PitchTracker {
 
     public static var defaultBufferSize: UInt32 { 4_096 }
 
-    public init(sampleRate: Int32, hopSize: Double = Double(PitchTracker.defaultBufferSize), peakCount: Int = 20) {
+    public init(sampleRate: Double, hopSize: Double = Double(PitchTracker.defaultBufferSize), peakCount: Int = 20) {
         ptrack.size = hopSize
         ptrack.numpks = peakCount
-        ptrack.sr = Double(sampleRate)
+        ptrack.sr = sampleRate
         swift_zt_ptrack_init(p: &ptrack)
     }
